Reorganise rewrite packages: move replace and numberunit out of contrib (#1122)
- Rename querqy.rewrite.replace → querqy.rewrite.regexreplace
- Move querqy.rewrite.contrib.replace → querqy.rewrite.replace; move
  ReplaceRewriter and ReplaceRewriterFactory from contrib into that package
- Move querqy.rewrite.contrib.numberunit → querqy.rewrite.numberunit; move
  NumberUnitRewriter and NumberUnitRewriterFactory from contrib into that package
