--- querqy-core/src/main/java/querqy/rewrite/contrib/ReplaceRewriterFactory.java
@@ -1,57 +1,0 @@
-/*
- * SPDX-License-Identifier: Apache-2.0
- *
- * Copyright 2019 Querqy Contributors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package querqy.rewrite.contrib;
-
-import querqy.model.ExpandedQuery;
-import querqy.model.Term;
-import querqy.parser.QuerqyParser;
-import querqy.rewrite.QueryRewriter;
-import querqy.rewrite.RewriterFactory;
-import querqy.rewrite.SearchEngineRequestAdapter;
-import querqy.rewrite.contrib.replace.ReplaceInstruction;
-import querqy.trie.SequenceLookup;
-
-import java.io.IOException;
-import java.io.InputStreamReader;
-import java.util.Set;
-
-public class ReplaceRewriterFactory extends RewriterFactory {
-
-    private final SequenceLookup<ReplaceInstruction> sequenceLookup;
-    private final boolean ignoreCase;
-
-    public ReplaceRewriterFactory(final String id,
-                                  final InputStreamReader reader,
-                                  final boolean ignoreCase,
-                                  final String inputDelimiter,
-                                  final QuerqyParser querqyParser) throws IOException {
-        super(id);
-        this.ignoreCase = ignoreCase;
-        sequenceLookup = new querqy.rewrite.contrib.replace.ReplaceRewriterParser(reader, this.ignoreCase, inputDelimiter, querqyParser).parseConfig();
-    }
-
-    @Override
-    public QueryRewriter createRewriter(ExpandedQuery input, SearchEngineRequestAdapter searchEngineRequestAdapter) {
-        return new ReplaceRewriter(sequenceLookup);
-    }
-
-    @Override
-    public Set<Term> getCacheableGenerableTerms() {
-        return QueryRewriter.EMPTY_GENERABLE_TERMS;
-    }
-}
