--- querqy-core/src/main/java/querqy/rewrite/contrib/numberunit/NumberUnitQueryCreator.java
@@ -1,78 +1,0 @@
-/*
- * SPDX-License-Identifier: Apache-2.0
- *
- * Copyright 2020 Querqy Contributors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package querqy.rewrite.contrib.numberunit;
-
-import querqy.model.BooleanParent;
-import querqy.model.BoostQuery;
-import querqy.model.QuerqyQuery;
-import querqy.rewrite.contrib.numberunit.model.LinearFunction;
-import querqy.rewrite.contrib.numberunit.model.PerUnitNumberUnitDefinition;
-
-import java.math.BigDecimal;
-import java.math.RoundingMode;
-import java.util.List;
-
-public abstract class NumberUnitQueryCreator {
-
-    private final int scale;
-    private static final RoundingMode ROUNDING_MODE = RoundingMode.HALF_UP;
-
-    private final BigDecimal n100 = new BigDecimal(100);
-    private final BigDecimal n0 = new BigDecimal(0);
-
-    protected NumberUnitQueryCreator(int scale) {
-        this.scale = scale;
-    }
-
-    public int getScale() {
-        return this.scale;
-    }
-
-    public RoundingMode getRoundingMode() {
-        return ROUNDING_MODE;
-    }
-
-    public LinearFunction createLinearFunctionParameters(final BigDecimal x1, final BigDecimal y1,
-                                                            final BigDecimal x2, final BigDecimal y2) {
-        final BigDecimal x = x1.subtract(x2);
-        final BigDecimal y = y1.subtract(y2);
-
-        final BigDecimal m = x.compareTo(n0) != 0 ? y.divide(x, this.scale, ROUNDING_MODE) : n0;
-        final BigDecimal b = y1.subtract(x1.multiply(m)).setScale(this.scale, ROUNDING_MODE);
-
-        return new LinearFunction(m, b);
-    }
-
-    public BigDecimal calculatePercentageChange(BigDecimal number, BigDecimal percentage) {
-        return number.multiply(percentage).divide(n100, this.scale, ROUNDING_MODE);
-    }
-
-    public BigDecimal subtractPercentage(BigDecimal number, BigDecimal percentage) {
-        return number.subtract(calculatePercentageChange(number, percentage)).setScale(this.scale, ROUNDING_MODE);
-    }
-
-    public BigDecimal addPercentage(BigDecimal number, BigDecimal percentage) {
-        return number.add(calculatePercentageChange(number, percentage)).setScale(this.scale, ROUNDING_MODE);
-    }
-
-    public abstract QuerqyQuery<BooleanParent> createFilterQuery(
-            BigDecimal value, List<PerUnitNumberUnitDefinition> perUnitNumberUnitDefinitions);
-
-    public abstract BoostQuery createBoostQuery(
-            BigDecimal value, List<PerUnitNumberUnitDefinition> perUnitNumberUnitDefinitions);
-}
