--- querqy-core/src/main/java/querqy/rewrite/contrib/replace/TermsReplaceInstruction.java
@@ -1,39 +1,0 @@
-/*
- * SPDX-License-Identifier: Apache-2.0
- *
- * Copyright 2020 Querqy Contributors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package querqy.rewrite.contrib.replace;
-
-import querqy.rewrite.logging.ActionLog;
-import querqy.rewrite.logging.MatchLog;
-
-import java.util.List;
-
-public class TermsReplaceInstruction extends ReplaceInstruction {
-
-    private final List<? extends CharSequence> replacementTerms;
-
-    public TermsReplaceInstruction(final List<? extends CharSequence> replacementTerms) {
-        this.replacementTerms = replacementTerms;
-    }
-
-    @Override
-    public void apply(final List<CharSequence> seq, final int start, final int exclusiveOffset,
-                      final CharSequence wildcardMatch, final List<ActionLog> actionLogs) {
-        removeTermFromSequence(seq, start, exclusiveOffset, replacementTerms, actionLogs, MatchLog.MatchType.EXACT);
-        seq.addAll(start, replacementTerms);
-    }
-}
