--- querqy-core/src/main/java/querqy/rewrite/replace/RegexReplaceRewriterFactory.java
@@ -1,46 +1,0 @@
-/*
- * SPDX-License-Identifier: Apache-2.0
- *
- * Copyright 2026 Querqy Contributors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package querqy.rewrite.replace;
-
-import querqy.model.ExpandedQuery;
-import querqy.rewrite.QueryRewriter;
-import querqy.rewrite.RewriterFactory;
-import querqy.rewrite.SearchEngineRequestAdapter;
-
-import java.io.IOException;
-import java.io.InputStreamReader;
-
-public class RegexReplaceRewriterFactory extends RewriterFactory {
-
-    private final RegexReplacing replacing;
-
-    public RegexReplaceRewriterFactory(final String rewriterId, final InputStreamReader reader,
-                                       final boolean ignoreCase) throws IOException {
-        super(rewriterId);
-
-        this.replacing = new RegexReplaceRewriterRulesParser(reader, ignoreCase).parserConfig();
-
-    }
-
-    @Override
-    public QueryRewriter createRewriter(final ExpandedQuery input,
-                                        final SearchEngineRequestAdapter searchEngineRequestAdapter) {
-
-        return new RegexReplaceRewriter(replacing);
-    }
-}
